Moved simple commands into 'simple' folder
